refactor: move poll-store files to it's own direcotry
