+++ apps/web/core/providers/stores/poll-store/poll-store-interface.ts
